requirements.txt
streamlit
pandas
supabase
requests
xlsxwriter
plotly
google-generativeai>=0.7.0